+++ src/sender/sender_perf_main.c
@@ -104,7 +104,7 @@ int perf_main(int argc, char **argv) {
 				gru_logger_set_mininum(options->log_level);
 				break;
 			case 'p':
-				options->parallel_count = atoi(optarg);
+				options->parallel_count = (uint16_t) atoi(optarg);
 				break;
 			case 'd':
 				options->duration = gru_duration_from_minutes(atoi(optarg));
